% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.r
\name{clean_factors}
\alias{clean_factors}
\title{Convert All Factors in Dataframe to Strings}
\usage{
clean_factors(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a data.frame with factor columns converted to string
}
\description{
Takes a data.frame and converts any factor column
to strings.
}
