% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.r
\name{create_param_formatter}
\alias{create_param_formatter}
\title{Create Parameter Formatter}
\usage{
create_param_formatter(...)
}
\arguments{
\item{...}{arguments to be passed to format}
}
\value{
a function that takes a vector of numbers and returns a
formatted character vector
}
\description{
Create formatter function to be used for a set of parameters.
}
\examples{
formatter <- create_param_formatter(digits = 3)
formatter(mtcars$mpg)

}
