% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.r
\name{get_indefinite_article}
\alias{get_indefinite_article}
\title{Get Indefinite Article}
\usage{
get_indefinite_article(word)
}
\arguments{
\item{word}{the word to which the indefinite article is applied}
}
\value{
either 'a' or 'an' depending on the word
}
\description{
Get the correct indefinite article for a given word. Returns
'an' if word starts with starts with a vowel and 'a' otherwise.
}
\examples{
get_indefinite_article('fruit')
get_indefinite_article('apple')

}
