% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{Probability and Odds}
\alias{Probability and Odds}
\alias{prob_to_odds}
\title{Convert Between Probability & Odds}
\usage{
prob_to_odds(p)
}
\arguments{
\item{p}{a probability to convert to odds}

\item{odds}{odds to convert to probability}
}
\description{
Functions to convert between probabilities and odds.
}
