% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.r
\name{quoted_list_string}
\alias{quoted_list_string}
\title{Format array as quoted string list}
\usage{
quoted_list_string(x)
}
\arguments{
\item{x}{character vector of items to be quoted}
}
\value{
a character vector of length 1 containing a quoted comma-separated list
}
\description{
Takes a character array and returns a character vector of length 1
containing a quoted list of the input array.
}
\examples{
quoted_list_string(c('a','b','c'))

}
