% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{time_in_days}
\alias{time_in_days}
\title{Caclulate Time in Days}
\usage{
time_in_days(x, days_per_year)
}
\arguments{
\item{x}{Either "days", "weeks", "months", or "years".}

\item{days_per_year}{the number of days in a year to be used in conversions}
}
\value{
the numer of days in the unit of time
}
\description{
Takes a string representing a unit of time and
calculates the number of days contained per unit.
}
