% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.r
\name{to_list_item_output}
\alias{to_list_item_output}
\title{Format to Print as List Item}
\usage{
to_list_item_output(x, n = 6, skip = 0)
}
\arguments{
\item{x}{the object to be printed}

\item{n}{the number of levels to indent output}

\item{skip}{the number of lines of output to skip}
}
\value{
a character vector of output as list item
}
\description{
Takes an object and returns a character vector representing how
the object should be printed as a list item.
}
\examples{
to_list_item(mtcars)

}
